requirements.txt
absl-py==0.12.0
astunparse==1.6.3
bsuite==0.3.5
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
descartes==1.1.0
dm-env==1.4
dm-env-rpc==1.0.4
dm-memorytasks==1.0.2
dm-sonnet==2.0.0
dm-tree==0.1.5
docker==4.4.4
flatbuffers==1.12
frozendict==1.2
future==0.18.2
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphs==0.1.3
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.0
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
mizani==0.7.3
networkx==2.0.0
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
palettable==3.3.0
pandas==1.2.3
patsy==0.5.1
Pillow==7.2.0
plotnine==0.8.0
portpicker==1.3.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.1
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
statsmodels==0.12.2
svgwrite==1.4.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.3.17
Tree==0.2.4
trfl==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
